+++ Woof/WoofTests/NetworkLayer/DataModelTests/AllSittersResponseTests.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+final class AllSittersResponseTests: XCTestCase {
+    func testPetSittersPresentsInAPI() {
+        _ = AllSittersResponse(petSitters: nil).petSitters
+    }
+}